fix
Modify conditional check in CodecBitSetTest

Updated conditional check to compare byte array lengths for equality instead of inequality in the CodecBitSetTest. This change is due to an occasional zero concatenated at the end of the byte array, which leads to issues with the `toByteArray` method as `BitSet` is little endian.