Adjust EventListener usage in TemplateParser classes
This commit primary changes how the eventListener is set in the TemplateParser and TemplateParserCore classes. Previously, a null eventListener would be used, but now we use the JavaHelper's nonNullOrDefault function to set a default eventListener if a null one is given. This effectively removes any null values, reducing potential null point exception risks and improving program stability.