Optimize root object fetching in TemplateParser class
Updated several methods in the TemplateParser class to fetch the root object only once at the beginning of the method instead of repeatedly within the method. This efficiency improvement simplifies code and reduces unnecessary function calls. Additionally, restructured 'encodeField' method in ParserWriterHelper class to streamline the process of generating error messages in exceptions.