Update AnnotationDescriptor to store name instead of type
The AnnotationDescriptor class is updated to store the name of the binding type rather than the type itself. This change affects multiple methods within the class that previously stored the type of the binding. The implementation of this change involves accessing the name of the type using the getName() method.