Update annotation descriptors with class type values
Updated several annotation descriptors to use the class type values instead of their names. Also modified several instances of key-value pairs within different methods to make use of simplified name representations for classes. This provides a more standard and useful representation, and leads to a reduction in the length of generated JSON description for testing.