Add bind subtypes descriptor and update README
A new descriptor, BIND_SUBTYPES, has been added to detect and describe subtypes of user-defined classes. The README file has been updated accordingly with the inclusion of a new version change and additional notes. As part of this change, the Descriptor's describeRawMessage method visibility has also been updated to public.