Add method to retrieve text until terminator is found
New methods added to BitReaderData.java allow for retrieving text from a byte array until a designated termination point. The termination point is passed as a String object and then converted to a byte array for comparison. This feature can help in more efficient handling of data streams where termination points are necessary.