Refactor MinMaxDataValidator to improve code readability
This refactor separates validation logic into different methods to check the minimum compatibility and maximum compatibility of the value, rather than having these checks performed in 'validateDefaultAgainstMinAndMax' and 'validateDataAgainstMinAndMax'. The aim is to make the code cleaner and more maintainable. Now, the 'check' methods are being called with the value, its constraints, the configuration data and an indication of the value type.