Refactor ByteOrder enumeration and method locations to reduce dependencies
To streamline the code, we've relocated the ByteOrder enumeration and several other files along with it. We've also moved methods that deal with input types and value casting from the GenericHelper class to the ParserDataType class, further reducing dependencies. Major changes also include usage of the newly positioned ByteOrder class in different files and deletion of unnecessary codes from the BitSetHelper class.