Add custom exception messages and optimize code
Introduced specialized error messages in the CodecException class for the scenarios where a suitable codec could not be found. Errors now explicitly state if the issue is related to binding or an array of alternatives. Also replaced calls that created general CodecException with the new methods. Revised and simplified parts of the code, including the creation of an 'unhandled field' exception, removal of commented instructions and substituting a Dot constant with its String representation.