Extract cacheTrailingZeros as a method
The calculation for trailing zeros in cache has been decentralized into a separated function, cacheTrailingZeros(). This effort makes the code cleaner and easier to read, as it's now very clear what is the purpose of the computation. Furthermore, this extraction avoids duplication of the same computation in different parts of the code.