Add new behavior classes and make CodecHelper methods public
Added several new classes to control the behavior of other classes when dealing with different object types such as BitSet, Integer, String, etc. Increased the accessibility of methods in CodecHelper for these new classes to use. This update creates new features to facilitate reading and writing operations, and improves the code's modularity and versatility.