Refactor getChosenConverter method call in CodecObject and CommonBehavior
The getChosenConverter method call that was previously in the CodecHelper class has been moved to the CommonBehavior class for more efficient execution. This was done to centralize the conversion process for more coherent processing of behavior-related tasks. Some methods in the CommonBehavior class were also simplified for cleaner code execution.