Remove unnecessary JavaHelper class usage
The JavaHelper class usage was removed and replaced with direct null check operations to set the EventListener entities, which simplifies the code. Moreover, the codec loading process was also streamlined by removing unnecessary null check and directly accessing the codec, which will now throw an exception if the codec is null.