Refactor SpeciesInterface and functions to reduce type complexity
The type specification of SpeciesInterface and the related functions in various modules have been simplified. The redundancies in type pertaining to their usage in different files has been reduced by removing unnecessary type parameters. This makes the functions and interfaces more understandable and easy to use. It has also slightly decreased the complexity of the written code and improved its legibility.