Refactor to use BitReader more consistently across files
This commit improves consistency in handling BitReader across various test and source files. It removes all usage of full class path io.github.mtrevisan.boxon.core.helpers.BitReader and replaces it with a direct reference, as BitReader is already imported. It also includes changes to BitReaderData and the addition of a new test file (NonByteMultipleLengthsHex).