Remove unused import and update class accessibility
The BigInteger import which was not in use, has been removed from the TeltonikaHelper class. Moreover, the accessibility of the AVLData, GPSElement, IOElement classes in MessageHex has been changed from public to private. Furthermore, the redundant dataCount1 variable has been consolidated into the dataCount variable and the number of bits to skip has been adjusted accordingly.