Add ContextParameter annotation and usage
This update includes adding a new `ContextParameter` annotation which assigns a constant or calculated value to a parameter that will be added to the context before processing. The parameters include `name` and `value`. Furthermore, this annotation helps in adding context parameters across various classes in the Boxon project, improving the structure and functioning of templates. It can be noticed in the 'MessageHex.java' test file where it replaced redundancy of different sized property classes by just one `FixedSizeProperty` class through its context parameter 'valueSize'.