Add overwrite option to context parameter annotation
The commit introduces an overwrite option to the ContextParameter annotation, which allows or prevents overwriting of existing parameters in the context during evaluation. By default, overwriting is not allowed, and an exception is thrown when an attempt is made. A reference to this new feature has been added in the README file.