Add ContextParameter annotation to README
The commit adds a detailed description of the new `ContextParameter` annotation to the README file. This annotation assigns a constant or calculated value to a parameter that will be added to the context before processing. It's designed to be instantiated before field encoding or decoding, and removed promptly after the processing of the field is completed.