Add utility method for trimming and creating unmodifiable lists
A utility method has been added to the `JavaHelper` class which combines the operations of trimming an `ArrayList` to its current size and creating an unmodifiable version of it. This method is subsequently utilized in other classes to replace explicit calls to `ArrayList.trimToSize()` and `Collections.unmodifiableList()`, reducing code duplication and improving readability.