Remove generic type usage from decode method and handle map instances
The decode method has been updated to return 'Object' type instead of generic 'T'. This change required corresponding updates in the Function calls and typecasts in the tests. Maps are now treated distinctly from other objects when getting field values and setting field values, to manage variable size properties. A small enhancement in this commit includes updating the logger message with the template name instead of the class name.