Change return type to Object in decode methods
The decode methods in various classes including TemplateDecoder, TemplateParserInterface, and TemplateParser have been updated to return Object instead of a generic type <T>. As part of this change, the <T> parameter has also been removed from these methods. This update avoids returning a specific type, instead the actual type can be encapsulated in the returned Object and cast when necessary.