Add support for BindAsList annotation
The commit introduces support for the BindAsList annotation in the CodecObject class while generating more specific exception messages when handles for collection annotations are not found. Also, error handling in the ComposerThreadedTest class has been improved to check if the first response contains an error before getting the message. In the CodecDefault class, exceptions are thrown when collection annotations other than BindAsArray are found.